+++ lithops/utils.py
@@ -20,7 +20,6 @@
 import io
 import re
 import os
-import sys
 import pika
 import uuid
 import json
@@ -34,7 +33,7 @@ import logging.config
 import threading
 
 from lithops.storage.utils import create_job_key
-from lithops.constants import LOGGER_FORMAT, LOGGER_LEVEL
+from lithops.constants import LOGGER_FORMAT, LOGGER_LEVEL, LOGGER_STREAM
 
 logger = logging.getLogger(__name__)
 
@@ -128,17 +127,23 @@ def agg_data(data_strs):
     return b"".join(data_strs), ranges
 
 
-def setup_logger(log_level=LOGGER_LEVEL,
-                 stream=None,
-                 log_format=LOGGER_FORMAT):
-    """Setup default logging for lithops."""
+def setup_lithops_logger(log_level=LOGGER_LEVEL,
+                         log_format=LOGGER_FORMAT,
+                         stream=None, filename=None):
+    """Setup logging for lithops."""
+    if log_level is None:
+        return
+
     if stream is None:
-        stream = sys.stderr
+        stream = LOGGER_STREAM
+
+    if filename is None:
+        filename = os.devnull
 
     if type(log_level) is str:
         log_level = logging.getLevelName(log_level.upper())
 
-    logging.config.dictConfig({
+    config_dict = {
         'version': 1,
         'disable_existing_loggers': False,
         'formatters': {
@@ -147,21 +152,33 @@ def setup_logger(log_level=LOGGER_LEVEL,
             },
         },
         'handlers': {
-            'default': {
+            'console_handler': {
                 'level': log_level,
-                'class': 'logging.StreamHandler',
                 'formatter': 'standard',
+                'class': 'logging.StreamHandler',
                 'stream': stream
+            },
+            'file_handler': {
+                'level': log_level,
+                'formatter': 'standard',
+                'class': 'logging.FileHandler',
+                'filename': filename,
+                'mode': 'a',
             },
         },
         'loggers': {
             'lithops': {
-                'handlers': ['default'],
+                'handlers': ['console_handler'],
                 'level': log_level,
                 'propagate': False
             },
         }
-    })
+    }
+
+    if filename is not os.devnull:
+        config_dict['loggers']['lithops']['handlers'] = ['file_handler']
+
+    logging.config.dictConfig(config_dict)
 
 
 def create_handler_zip(dst_zip_location, entry_point_file, entry_point_name=None):